Task #139562 [Sprint 2]Create Request Leave UI
